feat: Make text_buffer more leaner, no more options and view for it
