+++ packages/ultron/src/unit_tests/test_text_buffer.rs
@@ -5,7 +5,7 @@ use nalgebra::Point2;
 #[test]
 fn line_length() {
     let raw = "Hello world";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     assert_eq!(buffer.total_lines(), 1);
     assert_eq!(buffer.line_width(0), 11);
 }
@@ -13,14 +13,14 @@ fn line_length() {
 #[test]
 fn test_get_char() {
     let raw = "Hello world";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     assert_eq!(buffer.get_char(6, 0), Some('w'));
 }
 
 #[test]
 fn test_get_text() {
     let raw = "Hello world";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     let txt = buffer.get_text(Point2::new(0, 0), Point2::new(4, 0));
     assert_eq!(txt, "Hello");
 }
@@ -28,7 +28,7 @@ fn test_get_text() {
 #[test]
 fn test_get_text_world() {
     let raw = "Hello world";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     let txt = buffer.get_text(Point2::new(6, 0), Point2::new(10, 0));
     assert_eq!(txt, "world");
 }
@@ -36,7 +36,7 @@ fn test_get_text_world() {
 #[test]
 fn test_get_text_multi_line_world() {
     let raw = "Hello\nworld and\neverywhere";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     let txt = buffer.get_text(Point2::new(0, 1), Point2::new(4, 1));
     assert_eq!(txt, "world");
 }
@@ -44,7 +44,7 @@ fn test_get_text_multi_line_world() {
 #[test]
 fn test_get_text_multi_line_and() {
     let raw = "Hello\nworld and\neverywhere";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     let txt = buffer.get_text(Point2::new(6, 1), Point2::new(8, 1));
     assert_eq!(txt, "and");
 }
@@ -52,7 +52,7 @@ fn test_get_text_multi_line_and() {
 #[test]
 fn test_get_text_multi_line_and_every() {
     let raw = "Hello\nworld and\neverywhere";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     let txt = buffer.get_text(Point2::new(6, 1), Point2::new(4, 2));
     assert_eq!(txt, "and\nevery");
 }
@@ -60,7 +60,7 @@ fn test_get_text_multi_line_and_every() {
 #[test]
 fn delete_last_char() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.delete_char(10, 0);
     assert_eq!(buffer.to_string(), "Hello worl");
 }
@@ -68,7 +68,7 @@ fn delete_last_char() {
 #[test]
 fn delete_first_char() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.delete_char(0, 0);
     assert_eq!(buffer.to_string(), "ello world");
 }
@@ -76,7 +76,7 @@ fn delete_first_char() {
 #[test]
 fn delete_5th_char() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.delete_char(5, 0);
     assert_eq!(buffer.to_string(), "Helloworld");
 }
@@ -84,7 +84,7 @@ fn delete_5th_char() {
 #[test]
 fn delete_in_2nd_line() {
     let raw = "Hello\nworld\nthere";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.delete_char(0, 2);
     assert_eq!(buffer.to_string(), "Hello\nworld\nhere");
 }
@@ -92,7 +92,7 @@ fn delete_in_2nd_line() {
 #[test]
 fn break_line() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(5, 0);
     assert_eq!(buffer.to_string(), "Hello\n world");
 }
@@ -100,7 +100,7 @@ fn break_line() {
 #[test]
 fn break_line_then_insert_1() {
     let raw = "Hello world\n\nHowdy";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(0, 1);
     assert_eq!(buffer.to_string(), "Hello world\n\n\nHowdy");
     buffer.insert_char(0, 1, '1');
@@ -110,7 +110,7 @@ fn break_line_then_insert_1() {
 #[test]
 fn break_line_in_non_existing_cell() {
     let raw = "Hello world\n\nHowdy";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(2, 1);
     assert_eq!(buffer.to_string(), "Hello world\n  \n\nHowdy");
     buffer.insert_char(0, 2, '1');
@@ -120,7 +120,7 @@ fn break_line_in_non_existing_cell() {
 #[test]
 fn break_line_then_insert_1_below() {
     let raw = "Hello world\n\nHowdy";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(0, 1);
     assert_eq!(buffer.to_string(), "Hello world\n\n\nHowdy");
     buffer.insert_char(0, 2, '1');
@@ -130,7 +130,7 @@ fn break_line_then_insert_1_below() {
 #[test]
 fn break_line_then_insert_2_below() {
     let raw = "Hello world\n\nHowdy";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(0, 1);
     buffer.break_line(0, 1);
     assert_eq!(buffer.to_string(), "Hello world\n\n\n\nHowdy");
@@ -141,28 +141,28 @@ fn break_line_then_insert_2_below() {
 #[test]
 fn join_line() {
     let raw = "Hello\n world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.join_line(0, 0);
     assert_eq!(buffer.to_string(), "Hello world");
 }
 
 #[test]
 fn ensure_line_exist5() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.ensure_line_exist(5);
     assert_eq!(buffer.total_lines(), 6);
 }
 
 #[test]
 fn ensure_cell_exist5_2() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.ensure_cell_exist(5, 2);
     assert_eq!(buffer.to_string(), "\n\n      ");
 }
 
 #[test]
 fn ensure_cell_exist0_0() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.ensure_cell_exist(0, 0);
     assert_eq!(buffer.to_string(), " ");
 }
@@ -170,7 +170,7 @@ fn ensure_cell_exist0_0() {
 #[test]
 fn insert_5_lines() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(5, 0);
     buffer.break_line(0, 1);
     buffer.break_line(0, 1);
@@ -182,7 +182,7 @@ fn insert_5_lines() {
 #[test]
 fn join_5_lines() {
     let raw = "Hello\n\n\n\n\n world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.join_line(0, 0);
     buffer.join_line(0, 0);
     buffer.join_line(0, 0);
@@ -194,7 +194,7 @@ fn join_5_lines() {
 #[test]
 fn insert_new_line_at_start() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(0, 0);
     assert_eq!(buffer.to_string(), "\nHello world");
 }
@@ -202,42 +202,42 @@ fn insert_new_line_at_start() {
 #[test]
 fn insert_new_line_at_end() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.break_line(11, 0);
     assert_eq!(buffer.to_string(), "Hello world\n");
 }
 
 #[test]
 fn insert_anywhere_col() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.insert_char(5, 0, 'Y');
     assert_eq!(buffer.to_string(), "     Y");
 }
 
 #[test]
 fn insert_anywhere_line() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.insert_char(0, 5, 'Y');
     assert_eq!(buffer.to_string(), "\n\n\n\n\nY");
 }
 
 #[test]
 fn insert_anywhere_cell() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.insert_char(2, 5, 'Y');
     assert_eq!(buffer.to_string(), "\n\n\n\n\n  Y");
 }
 
 #[test]
 fn insert_anywhere_cell_10_10() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.insert_char(10, 10, 'Y');
     assert_eq!(buffer.to_string(), "\n\n\n\n\n\n\n\n\n\n          Y");
 }
 
 #[test]
 fn breaking_line_anywhere_will_make_lines_on_it() {
-    let mut buffer = TextBuffer::from_str(Options::default(), "");
+    let mut buffer = TextBuffer::from_str("");
     buffer.break_line(2, 5);
     assert_eq!(buffer.to_string(), "\n\n\n\n\n  \n");
 }
@@ -245,7 +245,7 @@ fn breaking_line_anywhere_will_make_lines_on_it() {
 #[test]
 fn lines_2() {
     let raw = "Hello\nworld";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     assert_eq!(buffer.total_lines(), 2);
     assert_eq!(buffer.line_width(0), 5);
     assert_eq!(buffer.line_width(1), 5);
@@ -255,7 +255,7 @@ fn lines_2() {
 #[test]
 fn cjk() {
     let raw = "Hello 文件系统";
-    let buffer = TextBuffer::from_str(Options::default(), raw);
+    let buffer = TextBuffer::from_str(raw);
     assert_eq!(buffer.total_lines(), 1);
     assert_eq!(buffer.line_width(0), 14);
 }
@@ -263,7 +263,7 @@ fn cjk() {
 #[test]
 fn insert_end_cjk() {
     let raw = "Hello 文件系统";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(14, 0, 'Y');
     assert_eq!(buffer.to_string(), "Hello 文件系统Y");
 }
@@ -271,7 +271,7 @@ fn insert_end_cjk() {
 #[test]
 fn insert_end_cjk_same_insert_on_13th_or_14th() {
     let raw = "Hello 文件系统";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(14, 0, 'Y');
     assert_eq!(buffer.to_string(), "Hello 文件系统Y");
 }
@@ -279,7 +279,7 @@ fn insert_end_cjk_same_insert_on_13th_or_14th() {
 #[test]
 fn insert_end_cjk_but_not_15th() {
     let raw = "Hello 文件系统";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(15, 0, 'Y');
     assert_eq!(buffer.to_string(), "Hello 文件系统 Y");
 }
@@ -287,7 +287,7 @@ fn insert_end_cjk_but_not_15th() {
 #[test]
 fn replace_start() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.replace_char(0, 0, 'Y');
     assert_eq!(buffer.to_string(), "Yello");
 }
@@ -295,7 +295,7 @@ fn replace_start() {
 #[test]
 fn replace_middle() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.replace_char(2, 0, 'Y');
     assert_eq!(buffer.to_string(), "HeYlo");
 }
@@ -303,7 +303,7 @@ fn replace_middle() {
 #[test]
 fn replace_end() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.replace_char(4, 0, 'Y');
     assert_eq!(buffer.to_string(), "HellY");
 }
@@ -311,7 +311,7 @@ fn replace_end() {
 #[test]
 fn replace_char_on_next_page() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.replace_char(4, 30, 'Y');
     assert_eq!(buffer.to_string(), "Hello\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n    Y");
 }
@@ -319,7 +319,7 @@ fn replace_char_on_next_page() {
 #[test]
 fn insert_start() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(0, 0, 'Y');
     assert_eq!(buffer.to_string(), "YHello");
 }
@@ -327,7 +327,7 @@ fn insert_start() {
 #[test]
 fn insert_middle() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(2, 0, 'Y');
     assert_eq!(buffer.to_string(), "HeYllo");
 }
@@ -335,7 +335,7 @@ fn insert_middle() {
 #[test]
 fn insert_end() {
     let raw = "Hello";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_char(5, 0, 'Y');
     assert_eq!(buffer.to_string(), "HelloY");
 }
@@ -343,7 +343,7 @@ fn insert_end() {
 #[test]
 fn test_cut_text() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     let txt = buffer.cut_text(Point2::new(0, 0), Point2::new(4, 0));
     assert_eq!(txt, "Hello");
     assert_eq!(buffer.to_string(), " world");
@@ -352,7 +352,7 @@ fn test_cut_text() {
 #[test]
 fn test_cut_text_multi_line() {
     let raw = "before text\nHello world\nafter text";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     let txt = buffer.cut_text(Point2::new(0, 1), Point2::new(4, 1));
     assert_eq!(txt, "Hello");
     assert_eq!(buffer.to_string(), "before text\n world\nafter text");
@@ -367,7 +367,7 @@ fn test_cut_text_multi_line() {
 // text
 fn test_cut_text_2lines_multi_line() {
     let raw = "before text\nHello world\nafter text";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     let txt = buffer.cut_text(Point2::new(0, 1), Point2::new(4, 2));
     assert_eq!(txt, "Hello world\nafter");
     //FIXME: There should only be 1 \n here
@@ -388,13 +388,7 @@ fn test_cut_text_2lines_multi_line() {
 #[test]
 fn test_cut_text_2lines_multi_line_block_mode() {
     let raw = "before text\nHello world\nafter text";
-    let mut buffer = TextBuffer::from_str(
-        Options {
-            use_block_mode: true,
-            ..Default::default()
-        },
-        raw,
-    );
+    let mut buffer = TextBuffer::from_str(raw);
     let txt = buffer.cut_text(Point2::new(6, 1), Point2::new(10, 2));
     assert_eq!(txt, "world\ntext");
     assert_eq!(buffer.to_string(), "before text\nHello \nafter ");
@@ -403,7 +397,7 @@ fn test_cut_text_2lines_multi_line_block_mode() {
 #[test]
 fn test_insert_text() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_text(5, 0, "YYYY");
     assert_eq!(buffer.to_string(), "HelloYYYY world");
 }
@@ -411,7 +405,7 @@ fn test_insert_text() {
 #[test]
 fn test_insert_multi_line_text() {
     let raw = "Hello world";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_text(5, 0, "XXXX\nYYYY");
     assert_eq!(buffer.to_string(), "HelloXXXX\nYYYY world");
 }
@@ -419,7 +413,7 @@ fn test_insert_multi_line_text() {
 #[test]
 fn test_insert_multi_line_text_to_multi_line_text() {
     let raw = "before text\nHello world\nafter text";
-    let mut buffer = TextBuffer::from_str(Options::default(), raw);
+    let mut buffer = TextBuffer::from_str(raw);
     buffer.insert_text(5, 1, "XXXX\nYYYY");
     assert_eq!(
         buffer.to_string(),
